feat: add metric query command
